Added collaborators

+++ task/views.py
@@ -1,4 +1,5 @@
-from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
 
 from project.models import Project
@@ -8,10 +9,14 @@ from todolist.models import ToDoList
 
 @login_required()
 def add_task(request, project_id, todolist_id):
-    project = Project.objects.filter(created_by=request.user).get(pk=project_id)
-    todolist = ToDoList.objects.filter(created_by=request.user, project=project).get(pk=todolist_id)
+    project = get_object_or_404(Project, id=project_id)
+    todolist = ToDoList.objects.filter(project=project).get(pk=todolist_id)
 
     if request.method == 'POST':
+        if request.user != project.owner and request.user not in project.collaborators.all():
+            messages.error(request, "You do not have permission to access this project.")
+            return redirect('project:project_detail', pk=project.id)
+
         name = request.POST.get('name')
         description = request.POST.get('description')
         status = request.POST.get('status')
@@ -22,6 +27,8 @@ def add_task(request, project_id, todolist_id):
             status=status, created_by=created_by
         )
 
+        messages.success(request, 'Task added successfully')
+
         return redirect('todolist:todolist', project.id, todolist.id)
 
     return render(request, 'task/add_task.html', {'project': project, 'todolist': todolist})
@@ -29,20 +36,28 @@ def add_task(request, project_id, todolist_id):
 
 @login_required()
 def detail(request, project_id, todolist_id, pk):
-    project = Project.objects.filter(created_by=request.user).get(pk=project_id)
+    project = Project.objects.get(pk=project_id)
     todolist = ToDoList.objects.filter(project=project).get(pk=todolist_id)
     task = Task.objects.filter(project=project, todolist=todolist).get(pk=pk)
+
+    if request.user != project.owner and request.user not in project.collaborators.all():
+        messages.error(request, "You do not have permission to access this project.")
+        return redirect('project:project_detail', pk=project.id)
 
     return render(request, 'task/detail.html', {'project': project, 'todolist': todolist, 'task': task})
 
 
 @login_required()
 def edit(request, project_id, todolist_id, pk):
-    project = Project.objects.filter(created_by=request.user).get(pk=project_id)
+    project = Project.objects.get(pk=project_id)
     todolist = ToDoList.objects.filter(project=project).get(pk=todolist_id)
     task = Task.objects.filter(project=project, todolist=todolist).get(pk=pk)
 
     if request.method == 'POST':
+        if request.user != project.owner and request.user not in project.collaborators.all():
+            messages.error(request, "You do not have permission to edit this task.")
+            return redirect('project:project_detail', pk=project.id)
+
         name = request.POST.get('name')
         description = request.POST.get('description')
         status = request.POST.get('status')
@@ -59,10 +74,15 @@ def edit(request, project_id, todolist_id, pk):
 
 @login_required()
 def delete(request, project_id, todolist_id, pk):
-    project = Project.objects.filter(created_by=request.user).get(pk=project_id)
+    project = Project.objects.get(pk=project_id)
     todolist = ToDoList.objects.filter(project=project).get(pk=todolist_id)
     task = Task.objects.filter(project=project, todolist=todolist).get(pk=pk)
 
+    if request.user != task.created_by:
+        messages.error(request, "You do not have permission to delete this task.")
+        return redirect('todolist:todolist', project.id, todolist.id)
+
     task.delete()
+    messages.success(request, 'Task deleted successfully')
 
     return redirect('todolist:todolist', project.id, todolist.id)
