+++ todolist/views.py
@@ -3,7 +3,7 @@ import json
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.http import JsonResponse
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 
 from project.models import Project
 from .models import ToDoList
@@ -11,17 +11,23 @@ from .models import ToDoList
 
 @login_required()
 def todolist(request, project_id, pk):
-    project = Project.objects.filter(created_by=request.user).get(pk=project_id)
-    todolist = ToDoList.objects.filter(created_by=request.user, project=project).get(pk=pk)
+    project = Project.objects.filter(pk=project_id).first()
+    todolist = ToDoList.objects.filter(project=project).get(pk=pk)
 
-    return render(request, 'todolist/todolist.html', {'project': project, 'todolist': todolist})
+    if request.user == project.owner or request.user in project.collaborators.all():
+        return render(request, 'todolist/todolist.html', {'project': project, 'todolist': todolist})
 
 
 @login_required
 def add(request, project_id):
-    project = Project.objects.filter(created_by=request.user).get(pk=project_id)
+    project = Project.objects.get(pk=project_id)
 
     if request.method == 'POST':
+
+        if request.user != project.owner and request.user not in project.collaborators.all():
+            messages.error(request, "You do not have permission to add a new To-Do List.")
+            return redirect('project:project_detail', pk=project_id)
+
         name = request.POST.get('name')
         description = request.POST.get('description')
 
@@ -30,33 +36,20 @@ def add(request, project_id):
                                     project_id=project_id)
             return redirect('project:project_detail', pk=project_id)
 
-    return render(request, 'todolist/add.html', {'project': project})
-
-
-def save_task_view(request):
-    project = Project.objects.filter(created_by=request.user).get(pk=request.POST.get('project_id'))
-    if request.method == 'POST':
-        # Extract data from POST request
-        data = json.loads(request.body)
-        task_name = data.get('name')
-        task_description = data.get('description', '')  # Assuming description is optional
-
-        # Create the task
-        task = ToDoList.objects.create(name=task_name, description=task_description, project_id=project.id)
-
-        # Return success response
-        return JsonResponse({'message': 'Task saved successfully!', 'task_id': task.id})
-
-    # Return error response if request method is not POST
-    return JsonResponse({'error': 'Method not allowed'}, status=405)
+    return redirect('project:project_detail', pk=project_id)
 
 
 @login_required()
 def edit(request, project_id, pk):
-    project = Project.objects.filter(created_by=request.user).get(pk=project_id)
-    todolist = ToDoList.objects.filter(created_by=request.user, project=project).get(pk=pk)
+    project = Project.objects.get(pk=project_id)
+    todolist = ToDoList.objects.filter(project=project).get(pk=pk)
 
     if request.method == 'POST':
+        if request.user != project.owner and request.user not in project.collaborators.all():
+            messages.error(request, "You do not have permission to edit this To-Do List.")
+
+            return redirect('project:project_detail', pk=project_id)
+
         name = request.POST.get('name')
         description = request.POST.get('description')
 
@@ -71,10 +64,20 @@ def edit(request, project_id, pk):
 
 @login_required()
 def delete(request, project_id, pk):
-    project = Project.objects.filter(created_by=request.user).get(pk=project_id)
+    project = Project.objects.get(pk=project_id)
     todolist = ToDoList.objects.filter(project=project).get(pk=pk)
 
-    todolist.delete()
-    messages.success(request, 'ToDo list deleted successfully')
+    if request.user != project.owner and request.user not in project.collaborators.all():
+        messages.error(request, "You do not have permission to delete this To-Do List.")
+
+        return redirect('project:project_detail', pk=project_id)
+
+    if request.user == todolist.created_by or request.user == project.owner:
+        todolist.delete()
+        messages.success(request, 'ToDo list deleted successfully')
+
+        return redirect('project:project_detail', pk=project_id)
+
+    messages.error(request, "You do not have permission to delete this To-Do List.")
 
     return redirect('project:project_detail', pk=project_id)
